+++ CHANGELOG.md
@@ -35,6 +35,10 @@ information message when the product have no remaining order.
 - Improve the design of the dashboards layouts.
 - Improve the design of the dashboards sidebars.
 
+### Removed
+
+- Remove unused route in teacher dashboard.
+
 ## [2.22.0] - 2023-04-20
 
 ### Added